requirements.txt
streamlit==1.27.2
ultralytics==8.0.196
opencv-python==4.8.1.78
pymupdf==1.23.5
pillow==10.0.1
pandas==2.1.1
numpy==1.24.3